Extract build orchestration from BakeryConfig
Move build_targets() and its retry helper out of config/config.py into
a new posit_bakery/image/build.py module. The function now accepts
targets, base_path, and build settings as explicit parameters instead
of reading from self.

The bake_plan_targets() method is removed — callers now use
BakePlan.from_image_targets() directly since the one-liner wrapper
added no value.

- config/config.py drops imports for BakePlan, python_on_whales,
  DockerException, time, and several error types
- cli/build.py calls build_targets() and BakePlan directly
- _merge_sequential_build_metadata_files() inlined into build_targets()
  since it was only called from there
